backend/budgetcalculator/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mssql-django==1.4
pyodbc==4.0.39
